+++ test/dummy/test/integration/root_switch_dropdown_test.rb
@@ -13,6 +13,8 @@ class RootSwitchDropdownTest < ActionDispatch::IntegrationTest
     assert_includes response.body, "admin@admin.com"
     assert_includes response.body, "Password"
     assert_includes response.body, 'data-theme="rounded"'
+    assert_includes response.body, "flat_pack/application"
+    assert_includes response.body, "@hotwired/turbo-rails"
     refute_includes response.body, "data-recording-studio-default-layout"
     refute_includes response.body, "mt-28"
     refute_includes response.body, "fixed inset-0"
@@ -34,6 +36,11 @@ class RootSwitchDropdownTest < ActionDispatch::IntegrationTest
     assert_response :success
     assert_includes response.body, workspace.name
     assert_select "body[data-recording-studio-default-layout='true']", count: 1
+    assert_includes response.body, "flat_pack/application"
+    assert_includes response.body, "@hotwired/turbo-rails"
+    assert_includes response.body, "flat-pack-page-nav"
+    refute_includes response.body, "Sign out"
+    refute_includes response.body, 'href="/users/sign_out"'
   end
 
   test "root switch page renders with the host default layout" do